+++ test/E2ETest/ApplicationFactoryTests.cs
@@ -1,12 +1,11 @@
 ﻿using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.Tye;
 using Test.Infrastructure;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.Tye.UnitTests
+namespace E2ETest
 {
     public class ApplicationFactoryTests
     {
