+++ test/E2ETest/ApplicationTests.cs
@@ -1,11 +1,12 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
+using Microsoft.Tye;
 using Test.Infrastructure;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.Tye.UnitTests
+namespace E2ETest
 {
     public class ApplicationTests
     {
